PSR_SaaS/requirements.txt
pandas==1.0.1
tensorflow==1.15.2
keras==2.3.1
pillow==7.1.1
flask==1.1.2
flask-sqlalchemy==2.4.1
SQLAlchemy==1.3.16
flask_bcrypt
flask_login
flask_wtf
flask_mysqldb